pyproject.toml
[project]
name = "fastembed-server"
version = "2.0.0"
description = "FastEmbed gRPC server with Qdrant integration and file streaming support"
authors = [
    {name = "FastEmbed Server Team"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "protobuf>=5.29.0",
    "grpcio>=1.69.0",
    "grpcio-tools>=1.69.0",
    "fastembed-gpu>=0.7.0",
    "numpy",
    "torch",
    'onnxruntime-gpu>=1.20.1; platform_system != "Darwin"',
    'onnxruntime; platform_system == "Darwin"',
    "qdrant-client>=1.7.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "aiofiles>=23.2.0",
    "pyzmq>=25.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true